Members and Admin view split
