+++ frontend/src/app/dashboard/[organization_id]/events/[event_id]/page.tsx
@@ -1,72 +1,21 @@
-import { getSessionTokenOrRedirect } from '@/app/utils';
-import {EventDetails} from '@/types';
-import AttendanceTable from "@/app/dashboard/[organization_id]/events/[event_id]/AttendanceTable";
-import {fetchMembership} from "@/lib/dataFetchers";
-import {Button} from "@/components/ui/button";
-import {Edit, Trash} from "lucide-react";
+import React from 'react';
+import AdminEventView from './AdminEventView';
+import MemberEventView from './MemberEventView';
+import { fetchOrganization, fetchMembership } from '@/lib/dataFetchers';
 
-
-async function getEventDetails(
-    organization_id: string,
-    event_id: string
-): Promise<EventDetails> {
-    const sessionToken = await getSessionTokenOrRedirect();
-
-    const res = await fetch(
-        `${process.env.NEXT_PUBLIC_RAILS_SERVER_URL}/organizations/${organization_id}/events/${event_id}`,
-        {
-            method: 'GET',
-            headers: {
-                Authorization: `Bearer ${sessionToken}`,
-                'Content-Type': 'application/json',
-            },
-            cache: 'no-cache',
-        }
-    );
-
-    if (!res.ok) {
-        const errorDetails = await res.json();
-        console.error('Failed to fetch event details', errorDetails);
-        throw new Error(`Failed to fetch event details: ${res.status}`);
-    }
-
-    const data = await res.json();
-    return data as EventDetails;
-}
-
-
-export default async function Page({ params }: { params: { organization_id: string, event_id: string } }) {
-    const eventDetails = await getEventDetails(params.organization_id, params.event_id);
+export default async function MembersPage({ params }: { params: { organization_id: string, event_id: string } }) {
+    const organization = await fetchOrganization(params.organization_id);
     const membership = await fetchMembership(params.organization_id);
+
     const admin_rights = membership.role === 'creator' || membership.role === 'manager';
 
     return (
-        <div >
-            {/* Event Header */}
-            <div className="flex justify-between items-center mb-4">
-                <div>
-                    <h2 className="text-2xl font-bold">{eventDetails.name}</h2>
-                    <p className="text-gray-600">{eventDetails.description}</p>
-                    <p>
-                        Date: <strong>{new Date(eventDetails.date).toLocaleDateString()}</strong>
-                    </p>
-                </div>
-                {admin_rights && (
-                    <div className="flex space-x-2">
-                        <Button variant="outline">
-                            <Edit className="h-4 w-4 mr-1" />
-                            Edit Event
-                        </Button>
-                        <Button variant="destructive">
-                            <Trash className="h-4 w-4 mr-1" />
-                            Delete Event
-                        </Button>
-                    </div>
-                )}
-            </div>
-
-            {/* Attendance Table */}
-            <AttendanceTable attendances={eventDetails.attendances} />
+        <div>
+            {admin_rights ? (
+                <AdminEventView organization_id={ params.organization_id } event_id={ params.event_id }/>
+            ) : (
+                <MemberEventView organization_id={ params.organization_id } event_id={ params.event_id }/>
+            )}
         </div>
     );
 }
